+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.1
 
+* New function: `reload_config`
 * New option: `urgent_on_bell`
 * Fix some memory leaks